# docs/swagger.yaml
basePath: /api/albums/feed
definitions:
  http.Error:
    properties:
      message:
        type: string
    type: object
...
  http.loginInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  http.loginResponse:
    properties:
      id:
        type: integer
    type: object
  http.logoutResponse:
    properties:
      status:
        type: string
    type: object
...
  http.signUpResponse:
    properties:
      id:
        type: integer
    type: object
...
  models.Date:
    properties:
      time.Time:
        type: string
    type: object
...
  models.Sex:
    enum:
    - M
    - F
    - O
    type: string
    x-enum-varnames:
    - Male
    - Female
    - Other
...
  models.User:
    properties:
      avatarSrc:
        type: string
      birthDate:
        $ref: '#/definitions/models.Date'
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      sex:
        $ref: '#/definitions/models.Sex'
      username:
        type: string
    type: object
...
host: localhost:4443
info:
  contact:
    email: yarik1448kuzmin@gmail.com
    name: Fluire API Support
  description: Server API for Fluire Streaming Service Application
  title: Fluire API
  version: 1.0.1
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Login account
      parameters:
      - description: username and password
        in: body
        name: userInput
        required: true
        schema:
          $ref: '#/definitions/http.loginInput'
      produces:
      - application/json
      responses:
        "200":
          description: User created
          schema:
            $ref: '#/definitions/http.loginResponse'
        "400":
          description: Incorrect input
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Sign In
      tags:
      - Auth
  /api/auth/logout:
    get:
      consumes:
      - application/json
      description: Logout account
      produces:
      - application/json
      responses:
        "200":
          description: User loged out
          schema:
            $ref: '#/definitions/http.logoutResponse'
        "400":
          description: Logout fail
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Log Out
      tags:
      - Auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Create account
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: User created
          schema:
            $ref: '#/definitions/http.signUpResponse'
        "400":
          description: Incorrect input
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Sign Up
      tags:
      - Auth
...
swagger: "2.0"